Errors resolved, project Done
