--- src/components/comparison section/Comparison.module.css
@@ -1,121 +1,0 @@
-.card {
-  background-color: var(--black-color);
-  box-shadow: 0px 17px 18px rgba(0, 0, 0, 0.25);
-  border: 0px;
-  outline: 1px solid white;
-  outline-offset: -7px;
-
-  transition: transform .5s ease-in-out;
-  transform: translateY(0%) translateX(0) scale(1) rotate(0);
-
-  @apply px-11 py-14 text-2xl font-thin relative
-}
-
-.cardVisible.wrong {
-  animation: slideInFromLeft 2s;
-}
-
-@keyframes slideInFromLeft {
-  0% {
-    transform: translateX(-50%);
-    opacity: 0;
-  }
-
-  100% {
-    transform: translateY(0%);
-    opacity: 1;
-  }
-}
-
-.cardVisible.good {
-  animation: slideInFromRight 2s;
-}
-
-@keyframes slideInFromRight {
-  0% {
-    transform: translateX(50%);
-    opacity: 0;
-  }
-
-  100% {
-    transform: translateY(0%);
-    opacity: 1;
-  }
-}
-
-
-.card:first-child {
-  font-family: 'Playfair Display', serif;
-}
-
-.bottomCorner,
-.topCorner {
-  content: " ";
-  display: block;
-  width: 45px;
-  height: 45px;
-  position: absolute;
-}
-
-.bottomCorner {
-  top: 96%;
-  left: 10px;
-  transform: translate(0%, -100%);
-  border-bottom: 1px solid white;
-  border-left: 1px solid white;
-}
-
-.topCorner {
-  top: 12px;
-  right: 10px;
-  transform: translate(0%, 0%);
-  border-top: 1px solid white;
-  border-right: 1px solid white;
-}
-
-.bottomCorner::before {
-  top: 100%;
-  left: 4px;
-  transform: translate(0%, -110%);
-  border-bottom: 1px solid white;
-  border-left: 1px solid white;
-}
-
-.topCorner::before,
-.bottomCorner::before {
-  content: " ";
-  display: block;
-  width: 35px;
-  height: 35px;
-  /* background-color: white; */
-  position: absolute;
-}
-
-.topCorner::before {
-  top: 100%;
-  right: 4px;
-  transform: translate(0%, -110%);
-  border-top: 1px solid white;
-  border-right: 1px solid white;
-}
-
-.list {
-  @apply text-left mt-4
-}
-
-.dpoint {
-  @apply text-left text-sm font-thin
-}
-
-.wrong .point {
-  list-style: url("../../images/listWrong.svg");
-}
-
-.good .point {
-  list-style: url("../../images/listCheck.svg");
-}
-
-.point div {
-  font-weight: 100;
-  @apply text-sm text-left font-thin text-[var(--white-color)]
-}
