+++ src/components/consultation/Consultation.module.css
@@ -1,0 +1,97 @@
+.container {
+  @apply w-full px-7 py-2 mx-auto flex flex-col text-left
+}
+
+.heading {
+  font-family: 'Playfair Display', serif;
+  @apply text-2xl font-normal text-left text-[var(--white-color)]
+}
+
+.smallHeading {
+  font-family: 'Playfair Display', serif;
+  @apply text-lg font-normal text-left text-[var(--white-color)]
+}
+
+.textThin {
+  @apply text-sm font-thin text-left text-[var(--white-color)]
+}
+
+.textStrong {
+  @apply text-sm font-bold text-left text-[var(--white-color)]
+}
+
+.listPoint {
+  @apply my-3 flex flex-row flex-nowrap text-left
+}
+
+.ball {
+  @apply mt-1 mr-4 relative w-fit
+}
+
+.listPoint.pointVisible {
+  animation: slideInFromLeft 1s;
+}
+
+@keyframes slideInFromLeft {
+  0% {
+    opacity: 0;
+  }
+
+  50% {
+    transform: translateX(-100%);
+    opacity: 0;
+  }
+
+  90% {
+    transform: translateX(5%);
+    opacity: 1;
+  }
+
+  100% {
+    transform: translateX(0%);
+    opacity: 1;
+  }
+}
+
+
+.bigBall {
+  position: relative;
+  z-index: 10;
+  width: 17px;
+  height: 17px;
+  background-color: white;
+  border-radius: 100%;
+  background: rgba(255, 255, 255, 0.01);
+  backdrop-filter: blur(15px);
+  transform: rotate(-180deg);
+}
+
+.smallBall {
+  position: absolute;
+  z-index: 0;
+  width: 7px;
+  height: 7px;
+  background: #FFFFFF;
+  border-radius: 100%;
+  z-index: 0;
+}
+
+.bottomRight {
+  left: 68%;
+  top: 57%;
+}
+
+.bottomLeft {
+  left: -10%;
+  top: 57%;
+}
+
+.topLeft {
+  left: -10%;
+  top: 0%;
+}
+
+.topRight {
+  left: 68%;
+  top: 0;
+}
