--- src/components/contact section/ContactForm.module.css
@@ -1,31 +1,0 @@
-.form {
-  transform-origin: 0% 0%;
-  transition: transform 1s ease-in-out;
-  transform: translateY(0%);
-  @apply flex flex-col items-center w-full px-4 mx-auto
-}
-
-.active {
-  transform: translateY(0%);
-}
-
-.unactive {
-  transform: translateY(-110%);
-}
-
-.activeDiv,
-.unactiveDiv {
-  transition: height 1s ease-in-out;
-}
-
-.activeDiv {
-  height: 100%;
-}
-
-.unactiveDiv {
-  height: 0;
-}
-
-.sendAlert {
-  @apply bg-green w-11/12 flex flex-row justify-between items-center px-4 py-3 rounded-2xl fixed bottom-2 left-2/4 -translate-x-2/4 z-50
-}
