+++ src/components/contact/ContactForm.module.css
@@ -1,0 +1,16 @@
+.form {
+  @apply flex flex-col items-center w-full px-4 mx-auto
+}
+
+.sendAlert,
+.errorAlert {
+  @apply w-11/12 font-bold flex flex-row justify-between items-center px-4 py-3 rounded-2xl fixed bottom-2 left-2/4 -translate-x-2/4 z-50
+}
+
+.sendAlert {
+  @apply bg-green
+}
+
+.errorAlert {
+  @apply bg-red
+}
