+++ src/components/divider/Divider.module.css
@@ -1,30 +1,4 @@
-.divider {
-  width: 100%;
-  height: fit-content;
-  margin: 35px 0;
-  text-align: center;
-  font-size: 28px;
-  position: relative;
-  height: 1px;
-  background: repeating-linear-gradient(to right,
-      transparent,
-      transparent 16px,
-      white 0px,
-      white 26px);
-}
-
-/* .divider::after {
-  position: absolute;
-  top: 0;
-  right: 0;
-  transform: translateY(-50%);
-  content: " ";
-  height: 1px;
-  width: 44%;
-  border-top: 6px dashed white;
-  background: red repeat-x;
-} */
-
+.divider,
 .dividerBlack {
   width: 100%;
   height: fit-content;
@@ -33,11 +7,40 @@
   font-size: 28px;
   position: relative;
   height: 1px;
-  background: repeating-linear-gradient(to right,
-      transparent,
-      transparent 16px,
-      black 0px,
-      black 26px);
+}
+
+.divider::after,
+.divider::before {
+  position: absolute;
+  top: 0;
+  right: 0;
+  transform: translateY(-50%);
+  content: " ";
+  height: 1px;
+  width: 40%;
+  border-top: 6px dashed white;
+}
+
+.divider::before {
+  top: 0;
+  left: 0;
+}
+
+.dividerBlack::before,
+.dividerBlack::after {
+  position: absolute;
+  top: 0;
+  right: 0;
+  transform: translateY(-50%);
+  content: " ";
+  height: 1px;
+  width: 40%;
+  border-top: 6px dashed black;
+}
+
+.dividerBlack::before {
+  top: 0;
+  left: 0;
 }
 
 .dividerIcon {
@@ -47,6 +50,5 @@
   width: 24px;
   height: 28px;
   transform: translate(-50%, -50%);
-  /* background-color: initial; */
   width: 17%;
 }
