--- src/components/free consultation section/FreeConsultation.module.css
@@ -1,21 +1,0 @@
-.iconTop {
-  background: url("../../images/star.svg");
-  width: 60px;
-  height: 60px;
-  position: relative;
-  top: -35px;
-  left: 0;
-  /* transform: translate(0, -50%); */
-  z-index: 0;
-}
-
-.badgeIcon {
-  background: url("../../images//badge.svg");
-  width: 24px;
-  height: 28px;
-  z-index: 10;
-  position: relative;
-  top: 0;
-  left: 50%;
-  transform: translate(-50%, 50%);
-}
