+++ src/components/navigation/Navigation.module.css
@@ -2,13 +2,11 @@
   display: block;
   height: 26px;
   width: 32px;
-  /* position: relative;
-  top: 0px;
-  left: 90%; */
   z-index: 10;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
+  cursor: pointer;
 }
 
 .burgerNavigation .line,
@@ -49,7 +47,6 @@
   justify-content: space-between;
   transform-origin: 0% 0%;
   transition: transform .5s ease-in-out;
-  /* animation: slideIn .5s; */
 }
 
 .menuItems {
