+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import MainPage from '@/components/main page/MainPage';
+import MainPage from '@/components/mainPage/MainPage';
 import { Inter } from 'next/font/google';
 import Head from 'next/head';
 import { useEffect, useRef, useState } from 'react';
